Move BoogieStmtListBuilder to a separate class
